Update charity.py
Preparation of the data
Looks at the features and how skewed they are
most of the features only include very low
income housing. Then the features are transformed
using a logarithmic transformation in order
to reduce the range of values caused by outliers.
Visualizes both. Visualizations can be found in README.